Ensure state hygiene test reads files as UTF-8
